% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{get_post_info}
\alias{get_post_info}
\title{Get headers information from a blogpost}
\usage{
get_post_info(post_df)
}
\arguments{
\item{post_df}{a one line data.frame a post name,
its url.}
}
\value{
A tibble with the blog post name and url and for each
section its header and a number, from 0 (title) to the number
of headers
}
\description{
Get headers information from a blogpost
}
