% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot.R
\name{shot_region}
\alias{shot_region}
\title{Webshot regions of a blog post}
\usage{
shot_region(df, path)
}
\arguments{
\item{df}{A data.frame with post_name, header and number variables.}

\item{path}{Either a full or partial path, where to put the screenshots.}
}
\value{
Nothing, it does produce screenshots in the path,
using the post_name and numbers to name files.
}
\description{
Webshot regions of a blog post
}
