% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webshot_prettyplease.R
\name{webshot_prettyplease}
\alias{webshot_prettyplease}
\title{Title}
\usage{
webshot_prettyplease(url, seed = sum(utf8ToInt(url)))
}
\arguments{
\item{url}{url to post}

\item{seed}{random seed}
}
\description{
Title
}
\examples{
webshot_prettyplease("https://itsalocke.com/blog/auto-deploying-documentation-better-change-tracking-of-artefacts/")
}
