% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webshot_prettyplease.R
\name{webshot_prettyplease}
\alias{webshot_prettyplease}
\title{Title}
\usage{
webshot_prettyplease(url, path)
}
\arguments{
\item{url}{url to post}
}
\description{
Title
}
\examples{
webshot_prettyplease("https://itsalocke.com/blog/how-to-maraaverickfy-a-blog-post-without-even-reading-it/",
                               path = "here.png")
}
